+++ CHANGELOG.md
@@ -5,6 +5,7 @@
     have been replaced by the `__call()` magic method;
 * now the `InstallShell::createVendorsLinks()` method writes a warning if the
     origin vendor directory does not exist;
+* added `hasPlugin()` and `removePlugins()` methods for the `TestCaseTrait`;
 * `ConsoleIntegrationTestCase`, `IntegrationTestCase` and `TestCase` classes now
     have the `$app` property, that contains the instance of the current
     application;